create css files for moving the rules for a clean architecture
