+++ course-subscription-app/src/App.jsx
@@ -11,9 +11,9 @@ import MyCourses from "./pages/MyCourses";
 import Login from "./pages/Login";
 import ProtectedRoute from "./components/ProtectedRoute";
 import Dashboard from "./pages/admin/Dashboard";
-import ManageUsers from "./pages/admin/ManageUsers";
+import ManageUsers from "./pages/admin/manageUsers/ManageUsers";
 import ManageSubscriptions from "./pages/admin/ManageSubscriptions";
-import ManageCourses from "./pages/admin/ManageCourses";
+import ManageCourses from "./pages/admin/manageCourses.css/ManageCourses";
 import Analytics from "./pages/admin/Analytics";
 import Register from "./pages/Register";
 import { useCart } from "./store/CartContext";
