+++ course-subscription-app/src/pages/admin/manageUsers/manageUsers.css
@@ -1,0 +1,5 @@
+/* Table cells rules */
+.table-style .table-cell-style{
+    font-size: 16px;
+    color: #FFFFFF;
+  }
